Review replied - add modify panel insead of modify page. Adapt to the overall appearance of the application
