+++ frontend/web/src/App.tsx
@@ -3,7 +3,6 @@ import { HomePage } from './pages/HomePage';
 import { TableQRPage } from './pages/TableQRPage'
 import { ForecastPage } from './pages/ForecastPage';
 import { RestaurantListPage } from './pages/RestaurantsListPage';
-import { RestaurantModifyPage } from './pages/RestaurantModifyPage';
 import { Routes, Route } from 'react-router-dom';
 import Root from './Root';
 import { LoginPage } from './pages/LoginPage';
@@ -29,8 +28,7 @@ function App() {
               <Route path="/" element={<HomePage />} />
               <Route path="/qr" element={<ProtectedRoute requiredRoles={["ADMIN", "MANAGER", "OWNER"]}><TableQRPage /></ProtectedRoute>} />
               <Route path="/forecast" element={<ProtectedRoute requiredRoles={["ADMIN", "MANAGER", "OWNER"]}><ForecastPage /></ProtectedRoute>} />
-              <Route path="/restaurants" element={<ProtectedRoute requiredRoles={["ADMIN", "MANAGER", "OWNER"]}><RestaurantListPage /></ProtectedRoute>} />
-              <Route path="/restaurants/modify/:restaurantId" element={<ProtectedRoute requiredRoles={["ADMIN", "MANAGER", "OWNER"]}><RestaurantModifyPage /></ProtectedRoute>} />
+              <Route path="/restaurants" element={<ProtectedRoute requiredRoles={["ADMIN"]}><RestaurantListPage /></ProtectedRoute>} />
             </Route>
           </Routes>
         </AuthProvider>
